+++ CHANGELOG.md
@@ -5,6 +5,7 @@
  - Add ability to invert PWM duty cycle ([#444](https://github.com/kizniche/mycodo/issues/444))
  - Add ability to select landing page ([#444](https://github.com/kizniche/mycodo/issues/444))
  - Add ability to set setpoint from PID dashboard elements ([#444](https://github.com/kizniche/mycodo/issues/444))
+ - Add Conditional Actions: Activate/Deactivate Timer ([#440](https://github.com/kizniche/mycodo/issues/440))
 
 ### Bugfixes